tests/requirements.txt
pytest>=7.0.0
requests>=2.28.0
python-dotenv>=0.19.0
selenium>=4.0.0
psycopg2-binary>=2.9.0
